# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Essa é uma API de testes criada para testes usando a lingagem go",
        "title": "RestFUL API usando Go, Gin, Swagger e GORM + SQLServer",
        "contact": {
            "name": "Opus127",
            "url": "http://www.opus127.com.br",
            "email": "suporte@opus127.com.br"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/clientes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Busca todos os clientes",
                "operationId": "get-all-clientes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Cliente"
                        }
                    }
                }
            }
        },
        "/clientes/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Encontra um cliente",
                "operationId": "find-customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cliente ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Cliente"
                        }
                    }
                }
            }
        },
        "/customers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get all the customers",
                "operationId": "get-all-customers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Customer"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new customer",
                "operationId": "create-customer",
                "parameters": [
                    {
                        "description": "customer data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Customer"
                        }
                    }
                }
            }
        },
        "/customers/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Find a customer",
                "operationId": "find-customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Customer"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a customer",
                "operationId": "delete-customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Customer"
                        }
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "summary": "Update a customer",
                "operationId": "update-customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "customer data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Customer"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Cliente": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Customer": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}